c/types: moved topic_table_delta to types.h
