+++ src/v/cluster/types.h
@@ -312,6 +312,30 @@ public:
 private:
     ss::sstring _msg;
 };
+// delta propagated to backend
+struct topic_table_delta {
+    enum class op_type { add, del, update, update_finished, update_properties };
+
+    topic_table_delta(
+      model::ntp,
+      cluster::partition_assignment,
+      model::offset,
+      op_type,
+      std::optional<partition_assignment> = std::nullopt);
+
+    model::ntp ntp;
+    cluster::partition_assignment new_assignment;
+    model::offset offset;
+    op_type type;
+    std::optional<partition_assignment> previous_assignment;
+
+    model::topic_namespace_view tp_ns() const {
+        return model::topic_namespace_view(ntp);
+    }
+
+    friend std::ostream& operator<<(std::ostream&, const topic_table_delta&);
+    friend std::ostream& operator<<(std::ostream&, const op_type&);
+};
 
 struct create_acls_cmd_data {
     static constexpr int8_t current_version = 1;
